requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-bdd==7.0.1
pytest-playwright==0.4.3
playwright==1.40.0
httpx==0.25.2
allure-pytest==2.13.2
requests==2.31.0
pytest-xdist==3.5.0
